added Userlogout,jwt functions,blacklisting , restricted access

+++ resources/item.py
@@ -1,6 +1,9 @@
 
 from flask_restful import Resource,reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (jwt_required,get_jwt_claims,
+                                jwt_optional,get_jwt_identity,
+                                fresh_jwt_required
+                                )
 from models.item import ItemModel
 
 class Item(Resource):
@@ -16,16 +19,15 @@ class Item(Resource):
                         help="Store id is needed"
                         )
 
-    @jwt_required()
+    @jwt_required
     def get(self,name):
         item=ItemModel.find_by_item_name(name)
         if item:
             return item.json()
         return {"message":"Item not found"},404
 
+    @fresh_jwt_required
     def post(self,name):
-
-
         if ItemModel.find_by_item_name(name):
             return {'Message':"Item with name: '{}' already exits".format(name)},400
         request_data = Item.parser.parse_args()
@@ -37,7 +39,11 @@ class Item(Resource):
         else:
             return {"message":"Error occured while saving the item!"},401
 
+    @jwt_required
     def delete(self,name):
+        claims=get_jwt_claims()
+        if not claims['is_admin']:
+            return {"message":"Admin privilege required"},401
         item=ItemModel.find_by_item_name(name)
         if item:
             item.delete_from_db()
@@ -66,8 +72,15 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {"Items":[item.json() for item in  ItemModel.query.all()]}
+        user_id=get_jwt_identity()
+        items=[item.json() for item in  ItemModel.find_all()]
+        if user_id:
+            return {"items": items}, 200
+        return {"Items":[item['name'] for item in items],
+                "message":"Log in to get more data"
+                }, 200
